services/prbot/pyproject.toml
[project]
name = "prbot-api"
version = "1.0.0"
description = "PR Bot — workspace dispatch, preflight, and job-secrets API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "motor>=3.3.0",
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
    "PyJWT[crypto]>=2.8.0",
    "python-dotenv>=1.0.1",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-asyncio"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
